seems that WORD is defined wrong!!
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@1113 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/fileio/Cunlzexe.h
@@ -14,7 +14,7 @@
 
 #include <vector>
 
-typedef unsigned long WORD;
+typedef unsigned short WORD;
 typedef unsigned char BYTE;
 
 class Cunlzexe
